improve management of images
Create a dialog which can be opened by clicking
on the image of a resource. The dialog lists
a number of sizes and allows an admin to set the
URL of the image for that size, along with a
preview of the image.

Create a helper method for adding an overlay
over an image with a link to open the dialog,
along with a concern for controllers to edit
and update images from the dialog.

Turn this on for the following resources:
- Albums
- Artists
- Events
- Genres
- Songs