allow songs to be reordered
Create a Sortable concern (models and controllers) which allows
associations (`has_many :through`) to be sorted using the attribute
`position` on the join table.

Resources are reordered in the view using `sortable` from jQuery UI
which fires of an AJAX request after each sorting.

Playlists and Albums are made sortable.

This fixes #19.