Improve management of duplicates
Create a button, *Find duplicates* on `songs#show` which displays a
dialog with a search box. When songs are clicked in the search results
they are marked as duplicates.

Add the method `Duplicatable#duplicate_candidates` which returns a list
of possible duplicates of the given resource. Use this method to list
possible duplicates on `songs#show` for admins.

Add a new tab on `user#dashboard`, visible only for admins, where the
songs with the highest number of potential duplicates are listed.