Remove donations, translations and site_config
Major clean up. Remove all the stuff that isn’t needed any more:

### Donations
Just redirect to Paypal instead. Also rephrase the text on `/money` and
the donation section on `/`.

### Translations
We don’t have online translations yet. As it is half-done (not even
half) and of bad quality (it was created a long time ago), remove it
and perhaps reintroduce it later on in a proper feature branch.

### Site configurations
`Admin::Config` was actually only used for donations, so we remove it
completely from the app.

This fixes #38, #36.