Prevent duplicates from showing up in search hits
Filter out duplicates from being searched in Solr during search. This
caused pagination to be weird as the pagination took place before
duplicates were filtered out by our custom scope, and the total number
of hits included duplicates as well.

The solution is to add `archetype_id` as a field indexed by Solr (hard
set to 0 for non-duplicatable models) and scope the search to exclude
any hits with a non-zero value for that field.

This fixes #28.