Start restyling the website (#84)
New front end
=============

The front-end on the website is complex and hard to manage. In order to make it
easier, I am simplifying it and using existing technologies such as:
- Bootstrap
- AngularJS
- Bower (via rails-assets gems)

All HTML, Javascript and CSS is rewritten from scratch. The front end is severly
simplified by leveraging Bootstrap and AngularJS instead of having to construct
all the boilerplate by hand.

In order to further clean up the markup, HTML with ERB is being replaced with
HAML, and SCSS is replaced by SASS.

Javascript Tests
================

Tests are extended to cover Javascript now as well. The
following test setup is used:
- Runner: Teaspoon
- Framework: Mocha
- Driver: PhantomJS
- Assertions: Chai and Sinon

Run the tests using `rake teaspoon`. This task is included in the default task
and is thus run when `rake` is invoked.

Activity feed
=============

The new activity feed displays the following activities:
- New listens
- New shares
- New playlists
- Modified playlists (with a special view when songs are added)

It resembles a scaled down Facebook feed. It currently shows *all*
activities on the website but should in the future be filtered to only
show activities from users and playlists that the viewing user is
following.

The new device view is scaled down from before. This has been removed:
- Volume
- Currently playing song

Cache Avatars and Names
=======================

Avatars from external accounts are now cached and stored as URLs. As are names.
So if the user changes name or avatar at the connected account, it is no longer
reflected automatically but needs a manual update from the user. This increases
speed though, and minimizes traffic to backends.

Semantic URLs
=============

Use `FriendlyId` to improve URLs to resources.

- User profile URLs are now `/:username`.
- Playlist URLs are now `/:username/:playlist`.
- Artist URLs are now `/artists/:artist_name`.
- Genre URLs are now `/genres/:genre_name`.
- Album URLs are now `/albums/:album_name`.
- Event URLs are now `/events/:event_name`.

Other notable changes
=====================

- Upgrade all gems.
- Create a fake data generator: `rake faker`.
- Extend searches to include backends automatically when nothing is found in the
  local database.
- Create a music page which handles all `#index` pages for media resources.
- Rename locales to language codes (`:en`) instead of country code (`:us`).
- Simplify the `rankable` concern.
- Add `solr` configuration to version control.
- Add A/B testing for root page.
- Convert logo into a vector image.
- Start splitting up `ApplicationController`.
- Breadcrumbs on certain pages to improve navigation.
- Content of `flash[:notice]` and `flash[:success]` is now displayed as a Growl-
  like notification in the bottom left corner.
- Suggestions to names or titles for albums, artists, etc. can now be done by
  using the `c` parameter, such as `c=artists`. The default is `queries`.
- Disable redirection and combining of associations by setting `dedup=1` when
  calling `/songs/SONG.json`.
- Switch to YARD for generating documentation.

Work left
=========

- Apps
- Devices
- Favicons